Remove timeout
