+++ ci/environment.yml
@@ -5,7 +5,6 @@ dependencies:
   - jinja2
   - packaging
   - pytest
-  - pytest-timeout
   - pytest-xdist
   - python
   - pyyaml
